+++ R/comp_barplot.R
@@ -1,30 +1,48 @@
 #' Plot (grouped and ordered) compositional barplots
 #'
-#' Stacked barplots showing composition of phyloseq samples for a specified number of coloured taxa. `comp_barplot` performs the compositional transformation for you, so your phyloseq object should contain counts!
+#' @description
+#' Stacked barplots showing composition of phyloseq samples for a specified number of coloured taxa.
+#' Normally your phyloseq object should contain count data, as by default `comp_barplot()` performs the "compositional" taxa transformation for you, and requires count input for some sample_order methods!
+#'
 #' - sample_order: Either specify a list of sample names to order manually, or the bars/samples can/will be sorted by similarity, according to a specified distance measure (default aitchison),
 #' - seriate_method specifies a seriation/ordering algorithm (default Ward hierarchical clustering with optimal leaf ordering, see seriation::list_seriation_methods())
 #' - group_by: You can group the samples on distinct plots by levels of a variable in the phyloseq object. The list of ggplots produced can be arranged flexibly with the patchwork package functions. If you want to group by several variables you can create an interaction variable with interaction(var1, var2) in the phyloseq sample_data BEFORE using comp_barplot.
 #' - facet_by can allow faceting of your plot(s) by a grouping variable. Using this approach is less flexible than using group_by but means you don't have to arrange a list of plots yourself like with the group_by argument. Using facet_by is equivalent to adding a call to facet_wrap(facets = facet_by, scales = "free") to your plot(s). Calling facet_wrap() yourself is itself a more flexible option as you can add other arguments like the number of rows etc. However you must use keep_all_vars = TRUE if you will add faceting manually.
 #' - bar_width: No gaps between bars, unless you want them (decrease width argument to add gaps between bars).
-#' - bar_outline_colour: Bar outlines default to "black". Set to NA if you don't want outlines.
+#' - bar_outline_colour: Bar outlines default to "grey5" for almost black outlines. Use NA if you don't want outlines.
+#' - merge_other: controls whether bar outlines can be drawn around individual (lower abundance) taxa that are grouped in "other" category. If you want to see the diversity of taxa in "other" use merge_taxa = FALSE, or use TRUE if you prefer the cleaner merged look
 #' - palette: Default colouring is consistent across multiple plots if created with the group_by argument, and the defaults scheme retains the colouring of the most abundant taxa irrespective of n_taxa
+#'
+#' @details
 #'
 #' @param ps phyloseq object
 #' @param tax_level taxonomic aggregation level (from rank_names(ps))
 #' @param n_taxa how many taxa to colour show distinct colours for (all other taxa grouped into "Other").
 #' @param tax_order order of taxa within the bars, currently only "abundance" works, which puts the most abundant taxa at the bottom (or left).
+#' @param merge_other if FALSE, taxa coloured/filled as "other" remain distinct, and so can have bar outlines drawn around them
 #' @param taxon_renamer function that takes taxon names and returns modified names for legend
-#' @param palette palette for taxa fill colours
-#' @param sample_order vector of sample names, any distance measure in dist_calc that does not require a phylogenetic tree, or "default" for the order returned by phyloseq::sample_names(ps)
-#' @param order_with_all_taxa if TRUE, this will use all taxa (not just the top n_taxa) to calculate distances for sample ordering
-#' @param tax_transform_for_ordering transformation of taxa values used before ordering samples by similarity
-#' @param label could also consider arbitrary annotation with extra info, like in complex heatmap
-#' @param group_by splits dataset by this variable (must be categorical) - resulting in a list of plots, one for each level of the group_by variable.
-#' @param facet_by facets plots by this variable (must be categorical) - if group_by is also set, the faceting with occur separately in the plot for each group.
-#' @param bar_width default 1 avoids random gapping otherwise seen with many samples (set to something less than 1 to introduce gaps between fewer samples)
+#' @param sample_order
+#' vector of sample names;
+#' or any distance measure in dist_calc that does not require a phylogenetic tree;
+#' or "default" for the order returned by phyloseq::sample_names(ps)
+#' @param order_with_all_taxa
+#' if TRUE, this will use all taxa (not just the top n_taxa) to calculate distances for sample ordering
+#' @param label sample label variable name
+#' @param group_by splits dataset by this variable (must be categorical)
+#' - resulting in a list of plots, one for each level of the group_by variable.
+#' @param facet_by facets plots by this variable (must be categorical)
+#' - if group_by is also set, the faceting with occur separately in the plot for each group.
+#' @param bar_width default 1 avoids random gapping otherwise seen with many samples
+#' (set to something less than 1 to introduce gaps between fewer samples)
 #' @param bar_outline_colour line colour separating taxa and samples (use NA for none)
-#' @param keep_all_vars slows down processing but is required for future plot customisation options like faceting
+#' @param bar_outline_width width of line separating taxa and samples (for no outlines set bar_outline_colour = NA)
+#' @param palette palette for taxa fill colours
+#' @param tax_transform_for_ordering transformation of taxa values used before ordering samples by similarity
+#' @param tax_transform_for_plot default "compositional" draws proportions of total counts per sample,
+#' but you could reasonably use another transformation, e.g. "identity", if you have truly quantitative microbiome profiling data
 #' @param seriate_method name of any ordering method suitable for distance matrices (see ?seriation::seriate)
+#' @param keep_all_vars TRUE may slow down internal melting with ps_melt for large phyloseq objects
+#' but is required for some post-hoc plot customisation
 #'
 #' @return ggplot or list of harmonised ggplots
 #' @export
@@ -56,6 +74,23 @@
 #'     sample_order = "default",
 #'     bar_width = 0.7,
 #'     bar_outline_colour = "black",
+#'     order_with_all_taxa = TRUE,
+#'     label = "subject_timepoint"
+#'   ) + coord_flip()
+#'
+#' # how many taxa are in those light grey "other" bars?
+#' # set merge_other, to find out (& remember to set a bar_outline_colour)
+#' dietswap %>%
+#'   ps_mutate(subject_timepoint = interaction(subject, timepoint)) %>%
+#'   ps_filter(nationality == "AAM", group == "DI", sex == "female") %>%
+#'   ps_arrange(desc(subject), desc(timepoint)) %>%
+#'   comp_barplot(
+#'     tax_level = "Genus", n_taxa = 12,
+#'     sample_order = "default",
+#'     merge_other = FALSE,
+#'     bar_width = 0.7,
+#'     bar_outline_colour = "black",
+#'     order_with_all_taxa = TRUE,
 #'     label = "subject_timepoint"
 #'   ) + coord_flip()
 #'
@@ -87,7 +122,7 @@
 #' # make a list of 2 harmonised composition plots (grouped by sex)
 #' p <- comp_barplot(dietswap,
 #'   n_taxa = 15, tax_level = "Genus",
-#'   bar_outline_colour = "black",
+#'   bar_outline_colour = "black", merge_other = TRUE,
 #'   sample_order = "aitchison", group_by = "sex"
 #' )
 #'
@@ -112,19 +147,22 @@ comp_barplot <- function(
                          tax_level,
                          n_taxa = 8,
                          tax_order = "abundance",
+                         merge_other = TRUE,
                          taxon_renamer = function(x) identity(x),
-                         palette = c("lightgrey", rev(distinct_palette(n_taxa))),
                          sample_order = "aitchison",
                          order_with_all_taxa = FALSE,
-                         tax_transform_for_ordering = "identity",
                          label = "SAMPLE",
                          group_by = NA,
                          facet_by = NA,
                          bar_width = 1,
-                         bar_outline_colour = "black",
-                         group_other = FALSE,
-                         keep_all_vars = TRUE,
-                         seriate_method = "OLO_ward") {
+                         bar_outline_colour = "grey5",
+                         bar_outline_width = 0.05,
+                         palette =
+                           c("lightgrey", rev(distinct_palette(n_taxa))),
+                         tax_transform_for_ordering = "identity",
+                         tax_transform_for_plot = "compositional",
+                         seriate_method = "OLO_ward",
+                         keep_all_vars = TRUE) {
 
   # check phyloseq for common problems (and fix or message about this)
   ps <- phyloseq_validate(ps, remove_undetected = FALSE, verbose = TRUE)
@@ -136,12 +174,19 @@ comp_barplot <- function(
   # ps <- aggregate_top_taxa(ps, top = n_taxa, level = tax_level)
   ps <- tax_agg(ps, rank = tax_level, sort_by = sum, top_N = n_taxa)[["ps"]]
 
+  # merge "other" category into one taxon,
+  # to allow drawing lines, but not within "other"
+  if (isTRUE(merge_other)) {
+    ps <- tax_agg(ps, rank = "top", force = TRUE)[["ps"]]
+  }
+
   # set taxa order
   if (identical(tax_order, "abundance")) {
     # you want "other" to go first, then least abundant to most?
     ordered_taxa <- rev(unique(unclass(phyloseq::tax_table(ps))[, "unique"]))
   } else {
-    # e.g. should allow external taxa ordering from calling top_taxa on a larger superset phyloseq
+    # e.g. should allow external taxa ordering
+    # from calling top_taxa on a larger superset phyloseq
     ordered_taxa <- tax_order
   }
   # fix taxa colour scheme
@@ -149,7 +194,8 @@ comp_barplot <- function(
   if (identical(names(palette), NULL)) names(palette) <- ordered_top_taxa
 
   # determine sample ordering option
-  samples_ordered_by_similarity <- FALSE # default (may be overwritten with true)
+  # set default (which may then be overwritten with TRUE)
+  samples_ordered_by_similarity <- FALSE
   if (identical(sample_order, "default")) {
     ordered_samples <- phyloseq::sample_names(ps)
   } else if (length(sample_order) == 1) {
@@ -180,10 +226,17 @@ comp_barplot <- function(
     if (!identical(facet_by, NA)) {
       kept_vars <- c(union(kept_vars, facet_by))
     }
-    if (!samples_ordered_by_similarity && length(sample_order) == 1 && sample_order != "default") {
+    if (
+      !samples_ordered_by_similarity &&
+        length(sample_order) == 1 &&
+        !identical(sample_order, "default")
+    ) {
       kept_vars <- c(union(kept_vars, sample_order))
     }
-    phyloseq::sample_data(ps) <- data.frame(phyloseq::sample_data(ps), check.names = FALSE)[, kept_vars, drop = FALSE]
+    phyloseq::sample_data(ps) <- data.frame(
+      phyloseq::sample_data(ps),
+      check.names = FALSE
+    )[, kept_vars, drop = FALSE]
   }
 
   # define function to actually create the plot/plots
@@ -198,35 +251,37 @@ comp_barplot <- function(
       }
       ps_ordered <- ps_seriate(
         ps_ordered,
-        method = seriate_method, dist = sample_order, tax_transform = tax_transform_for_ordering
+        method = seriate_method, dist = sample_order,
+        tax_transform = tax_transform_for_ordering
       )
       ordered_samples <- phyloseq::sample_names(ps_ordered)
     }
+    ps <- tax_transform(ps, transformation = tax_transform_for_plot)[["ps"]]
     # create long dataframe from compositional phyloseq
-    df <- ps %>%
-      microbiome::transform(transform = "compositional") %>%
-      ps_melt()
+    df <- ps_melt(ps)
 
     # set abundance order of taxa at chosen rank level
-    df[["unique"]] <- factor(df[["unique"]], levels = ordered_taxa, ordered = TRUE)
-    df[["top"]] <- factor(df[["top"]], levels = ordered_top_taxa, ordered = TRUE)
-    # df[[tax_level]] <- factor(df[[tax_level]], levels = ordered_taxa, ordered = TRUE)
+    df[["unique"]] <-
+      factor(df[["unique"]], levels = ordered_taxa, ordered = TRUE)
+    df[["top"]] <-
+      factor(df[["top"]], levels = ordered_top_taxa, ordered = TRUE)
 
     # set sample order
-    df[["SAMPLE"]] <- factor(df[["Sample"]], levels = ordered_samples, ordered = TRUE)
+    df[["SAMPLE"]] <-
+      factor(df[["Sample"]], levels = ordered_samples, ordered = TRUE)
 
     # build plot
-    if (isTRUE(group_other)){
-      p <- ggplot2::ggplot(df, ggplot2::aes_string(x = "SAMPLE", y = "Abundance", fill = "top"))
-    } else {
-      p <- ggplot2::ggplot(df, ggplot2::aes_string(x = "SAMPLE", y = "Abundance", fill = "top", group = "Abundance"))
-    }
-    if (is.na(bar_outline_colour)) {
-      p <- p + ggplot2::geom_bar(position = "stack", stat = "identity", width = bar_width)
-    } else {
-      p <- p + ggplot2::geom_bar(position = "stack", stat = "identity", width = bar_width, colour = bar_outline_colour)
-    }
-
+    p <- ggplot2::ggplot(
+      data = df,
+      mapping = ggplot2::aes_string(
+        x = "SAMPLE", y = "Abundance", fill = "top"
+      )
+    )
+    p <- p + ggplot2::geom_col(
+      position = "stack", width = bar_width,
+      colour = bar_outline_colour, size = bar_outline_width
+    )
+    # theme plot
     p <- p +
       ggplot2::theme(
         panel.background = ggplot2::element_blank(),
@@ -239,7 +294,7 @@ comp_barplot <- function(
       ggplot2::scale_fill_manual(
         values = palette,
         labels = taxon_renamer,
-        guide = ggplot2::guide_legend(reverse = TRUE)
+        guide = ggplot2::guide_legend(title = tax_level, reverse = TRUE)
       )
 
     if (!identical(facet_by, NA)) {
@@ -252,7 +307,9 @@ comp_barplot <- function(
   # group_by / splitting entire phyloseq by the group_by variable
   group_by_var <- phyloseq::sample_data(ps)[[group_by]]
   if (anyNA(group_by_var)) {
-    message("Warning: replacing NAs with 'NA's in group_by variable: ", group_by)
+    message(
+      "Warning: replacing NAs with 'NA's in group_by variable: ", group_by
+    )
     phyloseq::sample_data(ps)[[group_by]][is.na(group_by_var)] <- "NA"
   }
   LEVELS <- unique(phyloseq::sample_data(ps)[[group_by]])
@@ -272,19 +329,4 @@ comp_barplot <- function(
     names(plots_list) <- LEVELS
     return(plots_list)
   }
-}
-
-# helper function, copied from microbiome::aggregate_top_taxa this function is marked for deprecation!
-aggregate_top_taxa <- function(x, top, level) {
-  # microbiome R package (microbiome.github.com)
-  # Copyright (C) 2011-2020 Leo Lahti, Sudarshan Shetty et al. <microbiome.github.io>
-  x <- microbiome::aggregate_taxa(x, level)
-  tops <- microbiome::top_taxa(x, top)
-  tax <- phyloseq::tax_table(x)
-  inds <- which(!rownames(tax) %in% tops)
-  tax[inds, level] <- "Other"
-  phyloseq::tax_table(x) <- tax
-  tt <- phyloseq::tax_table(x)[, level]
-  phyloseq::tax_table(x) <- phyloseq::tax_table(tt)
-  microbiome::aggregate_taxa(x, level)
 }